Module wobbly::rc

Single-threaded reference-counting pointers. ‘Rc’ stands for ‘Reference Counted’. This module provides extended functionality for std::rc.

The type Wobbly<T> provides wobbly-shared ownership of a value of type T, allocated on the heap. Like Weak, Wobbly is generally a non-owning (weak) pointer that can be upgraded to obtain an Rc, but this will return None if the value stored in the allocation has already been dropped. A Wobbly pointer can also be downgraded to obtain a Weak.

Unlike Weak, however, one or more Wobblys may also together share an owning (strong) pointer, which keeps the stored value alive. When the first Wobbly sharing an owning (strong) pointer is dropped, the owning (strong) pointer is released as well.

A Wobbly pointer can be created from an owned Rc using Wobbly::new, which consumes the owning (strong) pointer and creates an additional non-owning (weak) pointer as well. Invoking clone on Wobbly produces a new wobbly pointer pointer to same value, which shares the same one owning (strong) pointer. While all Wobblys produced by clone-ing are alive, they keep the value alive even if there are no other owned Rcs to it. Once the first of the Wobblys is dropped, the owned (strong) pointer is released, and if it was the only remaining owned (strong) pointer the value is dropped. Note that multiple calls to Wobbly::new create multiple independent groups of wobbly pointers which can all keep the value alive on their own.

A cycle between Wobbly pointers where all Wobblys that share the same owning (strong) pointer participate in the cycle will never be deallocated. However, the cycle can be broken by storing a clone of one of the cyclic Wobblys that can then be dropped.

Structs

Wobbly

A single-threaded reference-counting pointer, similar to Weak, which provides wobbly-shared ownership of a value of type T, allocated on the heap.